pyproject.toml
[tool.poetry]
name = "kf-backend-test"
version = "0.1.0"
description = "KF backend test"
authors = ["Jay <jay.veer1@protonmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "kf_backend_test"}]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "3.8.3"
aiosignal = "1.3.1"
anyio = "3.6.2"
async-timeout = "4.0.2"
attrs = "22.2.0"
certifi = "2022.12.7"
charset-normalizer = "2.1.1"
click = "8.1.3"
coverage = "7.1.0"
dnspython = "2.3.0"
email-validator = "1.3.1"
exceptiongroup = "1.1.0"
fastapi = "0.89.1"
frozenlist = "1.3.3"
h11 = "0.14.0"
httpcore = "0.16.3"
httptools = "0.5.0"
httpx = "0.23.3"
idna = "3.4"
iniconfig = "2.0.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
markupsafe = "2.1.2"
multidict = "6.0.4"
orjson = "3.8.5"
packaging = "23.0"
pluggy = "1.0.0"
pydantic = "1.10.4"
pytest = "7.2.1"
pytest-asyncio = "0.20.3"
pytest-cov = "4.0.0"
python-dateutil = "2.8.2"
python-dotenv = "0.21.1"
python-multipart = "0.0.5"
pytz = "2022.7.1"
pyyaml = "6.0"
rfc3986 = "1.5.0"
six = "1.16.0"
sniffio = "1.3.0"
starlette = "0.22.0"
tomli = "2.0.1"
typing-extensions = "4.4.0"
ujson = "5.7.0"
uvicorn = "0.20.0"
uvloop = "0.17.0"
watchfiles = "0.18.1"
websockets = "10.4"
yarl = "1.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# [tool.poetry.scripts]
# start = "main"